HSEARCH-2293 Correct the code to only generate JBoss Logger 3.1 compatible bytecode

+++ backends/jgroups/src/main/java/org/hibernate/search/backend/jgroups/impl/JGroupsMasterMessageListener.java
@@ -81,7 +81,7 @@ public class JGroupsMasterMessageListener implements Receiver {
 			if ( log.isDebugEnabled() ) {
 				log.debugf(
 						"There are %d Lucene docs received from slave node %s to be processed by master",
-						queue.size(),
+						(Integer) queue.size(),
 						message.getSrc()
 				);
 			}
